requirements.txt
fpdf2==2.8.3
pygame==2.6.1
PyQt5==5.15.11
PyQt5_sip==12.17.0
